Add a sendmail entry to the Core group.
At the moment, packages in the Core group (cronie, for example) require
a MTA. Without specifying a group, one is pulled in via yum depsolving
rules, usually 'shortest-name-wins'. That's a bad way to set a default,
and causes other side effects, like minimal installs requiring multiple
CDs.

I'm not convinced that sendmail is *better* than other alternatives;
however, it is consistent with the current defaults in Base and Mail
Server.

The 'regression' case would be someone doing inidividual package
selection deselecting sendmail, and selecting some other MTA. They'd get
both installed. Kickstart can be used to overcome this.

If we get to the point that things in the Core group no longer require a
MTA, we can remove this. Until then, we should not be pulling in a random
one.