Move recordmydesktop from abrt-desktop's dep chain to comps
This complements the libreport change to drop the hard Requires:
recordmydesktop. Instead, install recordmydesktop through comps so that
individual spins could choose if they want recordmydesktop shipped on
the live media / installed by default.

recordmydesktop used to be pulled in through the abrt-desktop ->
libreport -> recordmydesktop dep chain. I'm now adding it to comps for
all the spins that previously pulled it in through the abrt-desktop dep
chain ...

... except for the Desktop spin. GNOME Shell already has built-in screen
recording functionality and recordmydesktop would duplicate that.

https://bugzilla.redhat.com/show_bug.cgi?id=959475
https://bugzilla.gnome.org/show_bug.cgi?id=696247