f19 and f20: clean up admin-tools group, use it for LXDE, Xfce and MATE
The 'admin-tools' group contains several utilities that would actually be
useful for many desktops, but it got lost in the shuffle between F17 and
F18. This commit drops gnome-packagekit from it (other desktops don't
necessarily want it, they may choose apper or yumex) and adds abrt-desktop
(which all the desktops were adding separately), and adds the group to the
LXDE, Xfce and MATE desktops: I am not yet sure whether the KDE, Cinnamon
or Sugar desktops would want it, and I'm pretty sure GNOME does not. For
F20 only, the commit also drops all the 'optional' stuff from the group,
as optional is pretty much a no-op now anyway.