workstation: Explode a few groups into workstation-product
Workstation's package selection is currently not very well structured.
It's loosely based on comps groups, but on top of that comes a large
number of package removals and additions in the kickstart files. This
makes the current setup difficult to understand and maintain.

This commit paves the way to moving package selection back into comps
and getting rid of additions/removals in spin-kickstarts.

A large number of cuts are currently done from the dial-up, firefox,
gnome-desktop, input-methods and standard groups. This commit explodes
the packages in those groups into the workstation-product group in order
to make it possible to do Workstation package selection (cuts) here in
comps.

The main motivation behind this change is to make Workstation netinstall
work. To get there, package selection has to be done wholly in comps.

https://bugzilla.redhat.com/show_bug.cgi?id=1134524